Remove blank space under back on search
Sadly, when you search for something and open the tweet detail, a random blank space appears between the header and the column-content. For some reason, the top: 30px line doesn't come into effect, so this makes the header take up the additional space. The change to icon positioning (making it a % instead of a set number) is to allow the header to be different heights while preserving the centred back arrow.